Recycle nbtree pages deleted during same VACUUM.
Maintain a simple array of metadata about pages that were deleted during
nbtree VACUUM's current btvacuumscan() call.  Use this metadata at the
end of btvacuumscan() to attempt to place newly deleted pages in the FSM
without further delay.  It might not yet be safe to place any of the
pages in the FSM by then (they may not be deemed recyclable), but we
have little to lose and plenty to gain by trying.  In practice there is
a very good chance that this will work out when vacuuming larger
indexes, where scanning the index naturally takes quite a while.

This commit doesn't change the page recycling invariants; it merely
improves the efficiency of page recycling within the confines of the
existing design.  Recycle safety is a part of nbtree's implementation of
what Lanin & Shasha call "the drain technique".  The design happens to
use transaction IDs (they're stored in deleted pages), but that in
itself doesn't affect when recycling becomes safe.  What matters is
whether or not other backends might be able to observe inconsistencies
in the tree structure (that they cannot detect and recover from by
moving right) -- not what's going in inside VACUUM.

Note that running a simple serial test case involving a large range
DELETE followed by a VACUUM VERBOSE will probably show that any newly
deleted nbtree pages are not yet reusable/recyclable.  This is expected
in the absence of even one concurrent XID allocation.  This happens
because deleted pages have a "safexid" value that comes from the _next_
full transaction ID at the point that they're physically deleted.  If
the _next_ transaction ID has yet to be allocated, then naturally it
can't be in the past (committed or aborted) relative to any possible
MVCC snapshot.

An important high-level goal of this commit (as well as related recent
commits e5d8a999 and 9f3665fb) is to make expensive deferred cleanup
operations in index AMs rare in general.  If index vacuuming frequently
depends on the next VACUUM operation finishing off work that the current
operation started, then we're left with a complicated and hard to
predict set of behaviors.  Index vacuuming with relatively predictable
and decoupled behavior will enable future work in vacuumlazy.c that
makes it possible to vacuum only a subset of indexes.  That will need to
be decided for each index based on the costs and benefits, which can
vary considerably even among indexes on the same table.

Author: Peter Geoghegan <pg@bowt.ie>
Reviewed-By: Masahiko Sawada <sawada.mshk@gmail.com>
Discussion: https://postgr.es/m/CAH2-Wzk76_P=67iUscb1UN44-gyZL-KgpsXbSxq_bdcMa7Q+wQ@mail.gmail.com