Remove tupgone special case from vacuumlazy.c.
Retry the call to heap_prune_page() for the buffer being pruned and
vacuumed in rare cases where there is disagreement between the first
heap_prune_page() call and VACUUM's HeapTupleSatisfiesVacuum() call.
This was possible when a concurrently aborting transaction rendered a
live tuple dead in the tiny window between each check.  As a result,
VACUUM's definition of dead tuples (tuples that are to be deleted from
indexes during VACUUM) is simplified: it is always LP_DEAD stub line
pointers from the first scan of the heap.  Note that in general VACUUM
may not have actually done all the pruning that rendered tuples LP_DEAD.

This decouples index vacuuming from initial heap scan's pruning, which
makes enables VACUUM to dynamically determine whether or not index
vacuuming should go ahead.  The index vacuum skipping performed by the
INDEX_CLEANUP mechanism added to Postgres 12 by commits a96c41f and
dd69597 worked this way already, but there are reasons to doubt that the
tupgone case is totally robust when INDEX_CLEANUP is enabled (it's
unclear at this point).  Simply retrying pruning is easy to understand
and clearly robust -- and so we are now on a much firmer footing to make
further improvements to index vacuuming.

Remove XLOG_HEAP2_CLEANUP_INFO records.  These are no longer necessary
because we now rely entirely on heap pruning to take care of recovery
conflicts during vacuum -- there is no question about not having a
conflict during an earlier pruning operation needed to make index scans
safe.

Note that heap vacuuming now uses exactly the same strategy for recovery
conflicts as index vacuuming.  This allows us to stop requiring a super
exclusive lock for the second scan of the heap (heap pruning), and to
completely avoid the need for recovery conflicts.  We split the records
used for heap pruning and heap vacuuming in two -- they're two rather
different kinds of operations now.

Bump XLOG_PAGE_MAGIC due to pruning and heap page vacuum WAL record
changes.