Aggressively advance OldestXmin during VACUUM.
This avoid issues with new_dead_tuples, a concern with order lines table
in particular.

It also makes it possible to set VM bits more often, since in practice
there are often new live tuples that are visible to all, despite not
being from before VACUUM's initial OldestXmin cutoff.  This is
commonplace, even.  At least it should be in a world with open and
closed pages.