Add page-level freezing to VACUUM.
Teach VACUUM to decide on whether or not to trigger freezing at the
level of whole heap pages.  Individual XIDs and MXIDs fields from tuple
headers now trigger freezing of whole pages, rather than independently
triggering freezing of each individual tuple header field.

Managing the cost of freezing over time now significantly influences
when and how VACUUM freezes.  Making freezing work at the level of whole
pages often results in writing less WAL over time.  Freezing all of a
given page's tuples together allows VACUUM to take full advantage of the
freeze plan WAL deduplication optimization added by commit 9e540599.

Also teach VACUUM to trigger page-level freezing whenever it detects
that heap pruning generated an FPI.  We'll have already written a large
amount of WAL just to do that much, so it's very likely a good idea to
get freezing out of the way for the page early.  This only happens in
cases where it will directly lead to marking the page all-frozen in the
visibility map.

In most cases "freezing a page" removes all XIDs < OldestXmin, and all
MXIDs < OldestMxact.  It doesn't quite work that way in certain rare
cases involving MultiXacts, though.  It is convenient to define "freeze
the page" in a way that gives FreezeMultiXactId the leeway to put off
the work of processing an individual tuple's xmax whenever it happens to
be a MultiXactId that would require an expensive second pass to process
aggressively (allocating a new multi is especially worth avoiding here).
FreezeMultiXactId is eager when processing is cheap (as it usually is),
and lazy in the event of an individual multi that happens to require
expensive processing.  This avoids regressions related to processing of
multis that page-level freezing would otherwise cause.

FreezeMultiXactId's no-op multi processing now "freezes" an xmax in the
event of an expensive-to-process xmax on a page when page-level freezing
triggers.  If, on the other hand, freezing is not triggered for the
page, then page-level no-op processing takes care of the multi for us
instead.  Either way, the remaining multi will ratchet back VACUUM's
relfrozenxid and/or relminmxid trackers to the extent required to make
everything safe, and we won't need an expensive second pass over the
multi (unless we really have no choice, for example during a VACUUM
FREEZE, where FreezeLimit always matches OldestXmin).

Later work will add eager freezing strategy to VACUUM (and reframe the
behavior established by this commit as lazy freezing, even though it's
not quite as lazy as the historical tuple-based approach to freezing).
Making freezing work at the page level is not just an optimization; it's
also a useful basis for modelling costs at the whole table level, since
it makes the visibility map a more reliable indicator of how far behind
(or ahead) we are on freezing at the level of the whole table.  Later
work that adds eager and lazy scanning strategies will build on this,
ultimately allowing VACUUM to advance relfrozenxid far more frequently.

Author: Peter Geoghegan <pg@bowt.ie>
Reviewed-By: Jeff Davis <pgsql@j-davis.com>
Reviewed-By: Andres Freund <andres@anarazel.de>
Discussion: https://postgr.es/m/CAH2-WzkFok_6EAHuK39GaW4FjEFQsY=3J0AAd6FXk93u-Xq3Fg@mail.gmail.com