Add page-level freezing to VACUUM.
Add a new strategy for freezing all-visible pages: freeze all of the
tuples on a page when a heap page is about to be marked all-visible.
This allows us to mark the page all-frozen in the VM early.  We also
support freezing driven entirely by FreezeLimit with the new approach.
Even there, we still freeze all the tuples on the page together, or none
at all.

For now only aggressive VACUUMs use the all-visible freezing strategy.
FreezeLimit-driven freezing is still useful for smaller tables.  The way
that we determine our freezing strategy will become more sophisticated
in a later commit.