Add all-visible FSM heuristic.
Have VACUUM treat all-visible pages differently when recording their
free space in the free space map during: record zero free space, unless
the true underlying amount of free space is too much to be worth putting
out of reach indefinitely.

Making all-visible pages resistant to change like this can be thought of as a
form of hysteresis.  The page is given an opportunity to "settle" and
permanently stay in the same state when the tuples on the page will never be
updated or deleted.  But when they are updated or deleted, the page can once
again be used to store any tuple.  Over time, most pages tend to settle
permanently in many workloads, perhaps only on the second or third attempt.