Teach VACUUM to use visibility map snapshot.
Also remove SKIP_PAGES_THRESHOLD behavior from vacuumlazy.c.  This makes
it very unlikely that non-aggressive VACUUMs will be able to advance
relfrozenxid -- all that it takes is one all-visible page to make that
impossible.  This will be addressed in a later commit that teaches
non-aggressive VACUUMs to be aware of (and to limit) the accumulation of
all-visible pages, and of the value in advancing relfrozenxid early.