Enhance pg_stat_progress_vacuum view.
Show the total number of scanned_pages for the VACUUM up front, and
repurpose the heap_blks_scanned column so that it shows the current
scanned_pages, not the current blkno.

Add a new column called heap_blk_processing to indicate which physical
heap block is currently being processed -- this is the column previously
known as heap_blks_scanned.  It still seems worth showing this, but it
doesn't seem useful to continue to count pages skipped using the
visibility map as either scanned or processed.

Finally, bring pg_stat_progress_vacuum in line with vacuumlazy.c by
renaming max_dead_tuples to max_dead_items, and num_dead_tuples to
num_dead_items.  We might as well do this now that we're breaking
compatibility with monitoring tools that worked on previous versions of
the view.