Avoid allocating MultiXacts during VACUUM.
Pass down vacuumlazy.c's OldestXmin cutoff to FreezeMultiXactId(), and
teach it the difference between OldestXmin and FreezeLimit.  Use this
high-level context to intelligently avoid allocating new MultiXactIds
during VACUUM operations.  We should always prefer to avoid allocating
new MultiXacts during VACUUM on general principle.  VACUUM is the only
mechanism that can claw back MultixactId space, so allowing VACUUM to
consume MultiXactId space (for any reason) adds to the risk that the
system will trigger the multiStopLimit wraparound protection mechanism.

FreezeMultiXactId() is now eager when it's cheap to process xmax, and
lazy when it's expensive/risky to process xmax (because an expensive
second pass that might result in allocating a new Multi is required).
We make a similar trade-off to the one made by vacuumlazy.c when a
cleanup lock isn't available on some heap page.  We can usually put off
freezing (for the time being) when it's inconvenient to proceed.  The
only downside to this approach is that it necessitates pushing back the
final relfrozenxid/relminmxid value that can be set in pg_class.