Finish removing aggressive mode VACUUM.
The concept of aggressive/scan_all VACUUM dates back to the introduction
of the visibility map in Postgres 8.4.  Although pre-visibility-map
VACUUM was far less efficient (especially after 9.6's commit fd31cd26),
its naive approach had one notable advantage: users only had to think
about a single kind of lazy vacuum (the only kind that existed).

Break the final remaining dependency on aggressive mode: replace the
rules governing when VACUUM will wait for a cleanup lock with a new set
of rules more attuned to the needs of the table.  With that last
dependency gone, there is no need for aggressive mode, so get rid of it.
Users once again only have to think about one kind of lazy vacuum.

In general, all of the behaviors associated with aggressive mode prior
to Postgres 16 have been retained; they just get applied selectively, on
a more dynamic timeline.  For example, the aforementioned change to
VACUUM's cleanup lock behavior retains the general idea of sometimes
waiting for a cleanup lock to make sure that older XIDs get frozen, so
that relfrozenxid can be advanced by a sufficient amount.  All that
really changes is the information driving VACUUM's decision on waiting.

We use new, dedicated cutoffs, rather than applying the FreezeLimit and
MultiXactCutoff used when deciding whether we should trigger freezing on
the basis of XID/MXID age.  These minimum fallback cutoffs (which are
called MinXid and MinMulti) are typically far older than the standard
FreezeLimit/MultiXactCutoff cutoffs.  VACUUM doesn't need an aggressive
mode to decide on whether to wait for a cleanup lock anymore; it can
decide everything at the level of individual heap pages.

It is okay to aggressively punt on waiting for a cleanup lock like this
because VACUUM now directly understands the importance of never falling
too far behind on the work of freezing physical heap pages at the level
of the whole table, following recent work to add VM scanning strategies.
It is generally safer for VACUUM to press on with freezing other heap
pages from the table instead.  Even if relfrozenxid can only be advanced
by relatively few XIDs as a consequence, VACUUM should have more than
ample opportunity to catch up next time, since there is bound to be no
more than a small number of problematic unfrozen pages left behind.
VACUUM now tends to consistently advance relfrozenxid (at least by some
small amount) all the time in larger tables, so all that has to happen
for relfrozenxid to fully catch up is for a few remaining unfrozen pages
to get frozen.  Since relfrozenxid is now considered to be no more than
a lagging indicator of freezing, and since relfrozenxid isn't used to
trigger freezing in the way that it once was, time is on our side.

Also teach VACUUM to wait for a short while for cleanup locks when doing
so has a decent chance of preserving its ability to advance relfrozenxid
up to FreezeLimit (and/or to advance relminmxid up to MultiXactCutoff).
As a result, VACUUM typically manages to advance relfrozenxid by just as
much as it would have had it promised to advance it up to FreezeLimit
(i.e. had it made the traditional aggressive VACUUM guarantee), even
when vacuuming a table that happens to have relatively many cleanup lock
conflicts affecting pages with older XIDs/MXIDs.  VACUUM thereby avoids
missing out on advancing relfrozenxid up to the traditional target
amount when it really can be avoided fairly easily, without promising to
do so (VACUUM only promises to advance up to MinXid/MinMulti).

XXX We also need to avoid the special auto-cancellation behavior for
antiwraparound autovacuums to make this truly safe.  See also, related
patch for this: https://commitfest.postgresql.org/41/4027/

Author: Peter Geoghegan <pg@bowt.ie>
Reviewed-By: Jeff Davis <pgsql@j-davis.com>
Reviewed-By: Andres Freund <andres@anarazel.de>
Discussion: https://postgr.es/m/CAH2-WzkU42GzrsHhL2BiC1QMhaVGmVdb5HR0_qczz0Gu2aSn=A@mail.gmail.com