Add skip scan to nbtree.
Skip scan allows nbtree index scans to efficiently use a composite index
on an index (a, b) for queries with a predicate such as "WHERE b = 5".
This is useful in cases where the total number of distinct values in the
column 'a' is reasonably small (think hundreds, possibly thousands).

In effect, a skip scan treats the composite index on (a, b) as if it was
a series of disjunct subindexes -- one subindex per distinct 'a' value.
We exhaustively "search every subindex" using a qual that behaves like
"WHERE a = ANY(<every possible 'a' value>) AND b = 5".

This approach is likely to be significantly less efficient than an index
scan that can use a single column index on 'b', of course.  But it can
also be orders of magnitude faster than a sequential scan.  We may very
well not have an index on "b", for whatever reason.

Author: Peter Geoghegan <pg@bowt.ie>