Stop conditioning nbtree actions on pin being held.
Commit 2ed5b87f, which taught nbtree to avoid holding onto a leaf page
buffer pin across the btgettuple calls for certain index scans, made it
possible for a scan position (either so->currPos or so->markPos) to be a
valid position without having a buffer pin held/a valid value in 'buf'.
The same commit introduced an idiom around buffer pin management: at the
point when a scan position was invalidated, a call to the function-like
macro BTScanPosUnpinIfPinned() released the buffer pin iff the relevant
position's 'buf' field was set to a non-InvalidBuffer value.  Experience
with bugfix commit 7c319f54 suggests that this idiom is rather brittle.

Finish off the work started by recent commit e6eed40e: fully move over
to conditioning releasing buffer pins on whether the scan in question is
a so->dropPin scan.  If it is, then _bt_drop_lock_and_maybe_pin must
have already dropped the buffer pin earlier on.  If it's not, then
_bt_drop_lock_and_maybe_pin certainly won't have dropped the pin when
called, so doing so is the responsibility of the pos-invalidating code.
BTScanPosUnpinIfPinned() is no longer needed, so get rid of it.  Keep
BTScanPosIsPinned(), but only call it from assertions going forward.

Now that the rules around buffer pins are a lot clearer, we can remove
all IncrBufferRefCount calls from nbtree.  They aren't truly necessary;
both of the IncrBufferRefCount call sites effectively just incremented
and then immediately decremented the backend-local pin count.  We can
safely avoid changing the buffer refcount by explicitly "transferring"
ownership of the pin from the old obsolescent position to the new one
that replaces it. (Actually, the btrestrpos caller _did_ keep around a
so->markPos/second pin, even after it returned, because the mark had to
be kept around in case it needed to be restored again some time later.
But we can safely keep such a mark around by setting so->markItemIndex,
and then invalidating the mark's original so->markPos representation.
This approach makes btrestrpos symmetrical with the approach taken by
_bt_steppage, which is the other former IncrBufferRefCount caller.)

This hardening/refactoring is a follow-up to bugfix commit 7c319f54.

Author: Peter Geoghegan <pg@bowt.ie>
Discussion: https://postgr.es/m/CAH2-Wzm3a8i31aBXi6oQt9uG7m1-xpX-MXjMMYoDJH=sBj1jnA@mail.gmail.com