Return TIDs in desc order during backwards scans.
Always return TIDs in descending order when returning a group of
duplicates to the scan whose TIDs come from an nbtree posting list
tuples during nbtree backwards scans.  This makes backwards scans tend
to require fewer buffer hits, since the scan is less likely to
repeatedly pin and unpin the same heap page/buffer (we'll get exactly as
many buffer hits as we get with a similar forwards scan case).

Commit 0d861bbb, which added nbtree deduplication, originally did things
this way to avoid interfering with _bt_killitems's previous approach to
setting LP_DEAD bits on posting list tuples.  _bt_killitems made a soft
assumption that it can iterate through posting lists in TID order,
finding corresponding killItems[]/so->currPos.items[] entries in that
same order (even when scanning backwards).  This worked out because of
the prior _bt_readpage backwards scan behavior. If we just change the
backwards scan posting list logic in _bt_readpage, and don't alter
_bt_killitems itself, we'll break its soft assumption.

Avoid that problem by converting the so->killedItems[] array into a
Bitmapset (that stores offsets into the so->currPos.items[] array) and
having _bt_killitems iterate through the Bitmapset.  That way the order
that items are originally saved in killItems can't matter.  There is an
additional benefit to using a Bitmapset for this: it is no longer
possible to run out of space in killedItems[] in cases where a
scrollable cursor repeatedly saves the same dead TID (from the same
page/so->currPos) by scrolling back and forth over it.  Calling
bms_add_member with the same item a second or a third time cannot make
_bt_killitems and less effective, since bms_add_member is idempotent.

Credit for the idea of using a Bitmapset for this goes to Tomas Vondra.

Author: Peter Geoghegan <pg@bowt.ie>
Reviewed-By: Mircea Cadariu <cadariu.mircea@gmail.com>
Discussion: https://postgr.es/m/CAH2-Wz=Wut2pKvbW-u3hJ_LXwsYeiXHiW8oN1GfbKPavcGo8Ow@mail.gmail.com