Return TIDs in desc order during backwards scans.
Always return items to the scan in descending key space order during
nbtree backwards scans.  In particular, always return TIDs in descending
order when returning a group of duplicates to the scan whose TIDs come
from an nbtree posting list tuples.  This makes nbtree backwards scans
tend to require fewer buffer hits, since the scan is less likely to
repeatedly pin and unpin the same heap page/buffer (we'll get exactly as
many buffer hits as we get with a similar forwards scan case).

Commit 0d861bbb, which added nbtree deduplication, originally did things
this way to avoid problems with how _bt_killitems sets LP_DEAD bits on
posting list tuples when scanning backwards.  We can work around that by
converting the so->killedItems[] array into a Bitmapset of int offsets
into the so->currPos.items[] array.

A Bitmapset is a natural choice for this.  It makes it easy for the call
to _bt_killitems (made when we step off a page) to iterate through
killedItems in ascending order (i.e. the order of so->currPos.items[]).
Recording the same dead item from the same page multiple times when a
scrollable cursor records the same dead item multiple times is now
idempotent.  It is no longer possible for us to run out of killedItems
space, either.

Credit for the idea of using a Bitmapset for this goes to Tomas Vondra.

Author: Peter Geoghegan <pg@bowt.ie>
Reviewed-By: Mircea Cadariu <cadariu.mircea@gmail.com>
Discussion: https://postgr.es/m/CAH2-Wz=Wut2pKvbW-u3hJ_LXwsYeiXHiW8oN1GfbKPavcGo8Ow@mail.gmail.com