Add amgetbatch interface.
Allows the index AM to provide items (TIDs and tuples) in per-leaf-page
batches.  A later commit will use this to implement I/O prefetching of
heap tuples in index scans (including index-only scans).  This is
similar to prefetching already done in bitmap scans, and can result in
significant speedups.

TODO: Add a new table AM interface that performs ordered index scans
that are directly under the control of the table AM, and avoid passing
around TIDs from the core executor to the table AM interface.

TODO: Make heapam use that new interface.

TODO: Add optimization that has heapam batch up accesses to the same
heap page during ordered index scans, such that we can avoid repeat
buffer locking and unlocking for groups of TIDs that point to the same
heap page.  This is possible in cases with index/heap correlation.