Add amgetbatch support to hash index AM.
This patch should be considered a work in progress.  It has only been
lightly tested, and it's not clear if I've accounted for all of the
intricacies with bucket splits (and with pins that are held by the
scan's opaque state more generally).

This automatically switched hash indexes over to using the dropPin
optimization, since that is standard when using the new amgetbatch
interface.  This won't bring similar benefits to hash index scans that
nbtree index scans gained when commit 2ed5b87f went in.  Hash index
vacuuming acquires a cleanup lock on bucket pages, but conflicting pins
on bucket pages are still held for the full duration of each index scan.

Note that the code path in _hash_kill_items that calls _hash_getbuf
(though only with the LH_OVERFLOW_PAGE flag/overflow pages) is dead
code/vestigial on master [1].  However, it's now useful again, since
_hash_kill_items might now need to relock and repin an overflow page
(not just relock it) during dropPin scans.

[1] https://postgr.es/m/CAH2-Wz=8mefy8QUcsnKLTePuy4tE8pdO+gSRQ9yQwUHoaeFTFw@mail.gmail.com